app/streamData/producer/requirements.txt
pandas
kafka-python